+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+- [2026-04-22] Gemini 3.1 flash image model (Nano Banana 2) is supported.
+
 - [2026-04-02] Switch to automatic caching for Claude 4.6 (but not for bedrock yet). Add message timestamp and round index to the tracer tool.
 
 - [2026-03-11] GPT-5.4 is supported. We now add `phase` labels to assistant messages, and preserve and send them to the server.